refactor(doctor): split configured plugin repair
