+++ src/commands/doctor/shared/missing-configured-plugin-install.repair.ts
@@ -1,0 +1,478 @@
+import { rm } from "node:fs/promises";
+import { stripAnsi } from "../../../../packages/terminal-core/src/ansi.js";
+import type { OpenClawConfig } from "../../../config/types.openclaw.js";
+import type { PluginInstallRecord } from "../../../config/types.plugins.js";
+import {
+  normalizeUpdateChannel,
+  resolveRegistryUpdateChannel,
+} from "../../../infra/update-channels.js";
+import type { ClawHubRiskAcknowledgementRequest } from "../../../plugins/clawhub.js";
+import {
+  loadInstalledPluginIndexInstallRecords,
+  writePersistedInstalledPluginIndexInstallRecords,
+} from "../../../plugins/installed-plugin-index-records.js";
+import { loadInstalledPluginIndex } from "../../../plugins/installed-plugin-index.js";
+import { loadManifestMetadataSnapshot } from "../../../plugins/manifest-contract-eligibility.js";
+import { updateNpmInstalledPlugins } from "../../../plugins/update.js";
+import { resolveUserPath } from "../../../utils.js";
+import { resolveCompatibilityHostVersion, VERSION } from "../../../version.js";
+import {
+  type BundledPluginPackageDescriptor,
+  collectConfiguredPluginIdsWithMissingChannelConfigDescriptors,
+  collectDownloadableInstallCandidates,
+  collectInstalledPluginIdsWithRepairablePackageDiagnostics,
+  collectInstalledPluginIdsWithStaleVersionBoundRuntimePackages,
+  collectOfficialReplacementInstallCandidates,
+  collectUpdateDeferredPluginIds,
+} from "./missing-configured-plugin-install.candidates.js";
+import {
+  collectBlockedPluginIds,
+  collectConfiguredChannelIds,
+  collectConfiguredPluginIds,
+  collectEffectiveConfiguredChannelOwnerPluginIds,
+} from "./missing-configured-plugin-install.ids.js";
+import {
+  appendClawHubRiskAcknowledgementGuidance,
+  installCandidate,
+  isActionableClawHubSkippedOutcome,
+  isClawHubReviewNotice,
+  recordClawHubInstallSpec,
+} from "./missing-configured-plugin-install.install.js";
+import {
+  forceNpmInstallRecordRepair,
+  isInstalledRecordMissingOnDisk,
+  isTrustedOfficialInstallRecordForCandidate,
+  pathsEqual,
+  recordMatchesBundledPackage,
+  resolveSafeBrokenOfficialInstallRemovalPath,
+} from "./missing-configured-plugin-install.records.js";
+import {
+  isLegacyPackageUpdateDoctorPass,
+  shouldDeferConfiguredPluginInstallRepair,
+} from "./update-phase.js";
+
+type RepairMissingPluginInstallsResult = {
+  /** User-facing repair notes for installed or recovered plugin records. */
+  changes: string[];
+  /** User-facing warnings for failed or skipped plugin install repairs. */
+  /** User-facing notices from successful repairs that still need operator review. */
+  notices?: string[];
+  warnings: string[];
+  /** Plugin ids successfully repaired from current configuration. */
+  repairedPluginIds?: string[];
+  /** User-facing details for repairs explicitly deferred until post-core convergence. */
+  deferredRepairDetails?: string[];
+  /** Plugin ids whose install repair failed and should be preserved from cleanup passes. */
+  failedPluginIds?: string[];
+  /**
+   * The full install-record map after repair. Equal to the input
+   * `baselineRecords` (or the disk-loaded records when no baseline was
+   * provided) plus any mutations (newly-installed payloads, removed stale
+   * bundled records). Callers that need to subsequently overwrite the
+   * persisted index MUST seed their write from this map — the disk has
+   * already been written to with the same set, but the in-memory caller
+   * state is stale otherwise.
+   */
+  records: Record<string, PluginInstallRecord>;
+};
+
+/** Repair missing installs inferred from the current OpenClaw config. */
+export async function repairMissingConfiguredPluginInstalls(params: {
+  cfg: OpenClawConfig;
+  env?: NodeJS.ProcessEnv;
+  acknowledgeClawHubRisk?: boolean;
+  onClawHubRisk?: (request: ClawHubRiskAcknowledgementRequest) => boolean | Promise<boolean>;
+  /**
+   * Optional pre-seeded records. When provided, this map is used instead of
+   * the disk-loaded install-record snapshot. Pass the in-memory records
+   * from earlier post-core steps (sync/npm) so this repair pass can layer
+   * its mutations on top of them rather than reading a stale disk
+   * snapshot. The merged result is persisted before this function returns.
+   */
+  baselineRecords?: Record<string, PluginInstallRecord>;
+}): Promise<RepairMissingPluginInstallsResult> {
+  return repairMissingPluginInstalls({
+    cfg: params.cfg,
+    env: params.env,
+    pluginIds: collectConfiguredPluginIds(params.cfg, params.env),
+    channelIds: collectConfiguredChannelIds(params.cfg, params.env),
+    blockedPluginIds: collectBlockedPluginIds(params.cfg),
+    ...(params.acknowledgeClawHubRisk ? { acknowledgeClawHubRisk: true } : {}),
+    ...(params.onClawHubRisk ? { onClawHubRisk: params.onClawHubRisk } : {}),
+    ...(params.baselineRecords ? { baselineRecords: params.baselineRecords } : {}),
+  });
+}
+
+/** Repair missing installs for an explicit plugin/channel id set. */
+export async function repairMissingPluginInstallsForIds(params: {
+  cfg: OpenClawConfig;
+  pluginIds: Iterable<string>;
+  channelIds?: Iterable<string>;
+  blockedPluginIds?: Iterable<string>;
+  env?: NodeJS.ProcessEnv;
+  baselineRecords?: Record<string, PluginInstallRecord>;
+  acknowledgeClawHubRisk?: boolean;
+  onClawHubRisk?: (request: ClawHubRiskAcknowledgementRequest) => boolean | Promise<boolean>;
+}): Promise<RepairMissingPluginInstallsResult> {
+  return repairMissingPluginInstalls({
+    cfg: params.cfg,
+    env: params.env,
+    pluginIds: new Set(
+      [...params.pluginIds].map((pluginId) => pluginId.trim()).filter((pluginId) => pluginId),
+    ),
+    channelIds: new Set(
+      [...(params.channelIds ?? [])]
+        .map((channelId) => channelId.trim())
+        .filter((channelId) => channelId),
+    ),
+    blockedPluginIds: new Set(
+      [...(params.blockedPluginIds ?? [])]
+        .map((pluginId) => pluginId.trim())
+        .filter((pluginId) => pluginId),
+    ),
+    ...(params.acknowledgeClawHubRisk ? { acknowledgeClawHubRisk: true } : {}),
+    ...(params.onClawHubRisk ? { onClawHubRisk: params.onClawHubRisk } : {}),
+    ...(params.baselineRecords ? { baselineRecords: params.baselineRecords } : {}),
+  });
+}
+
+async function repairMissingPluginInstalls(params: {
+  cfg: OpenClawConfig;
+  pluginIds: ReadonlySet<string>;
+  channelIds: ReadonlySet<string>;
+  blockedPluginIds?: ReadonlySet<string>;
+  env?: NodeJS.ProcessEnv;
+  baselineRecords?: Record<string, PluginInstallRecord>;
+  acknowledgeClawHubRisk?: boolean;
+  onClawHubRisk?: (request: ClawHubRiskAcknowledgementRequest) => boolean | Promise<boolean>;
+}): Promise<RepairMissingPluginInstallsResult> {
+  const env = params.env ?? process.env;
+  const snapshot = loadManifestMetadataSnapshot({
+    config: params.cfg,
+    env,
+  });
+  const currentBundledPlugins = loadInstalledPluginIndex({
+    config: params.cfg,
+    env,
+    installRecords: {},
+  }).plugins.filter((plugin) => plugin.origin === "bundled");
+  const knownIds = new Set([
+    ...snapshot.plugins.map((plugin) => plugin.id),
+    ...currentBundledPlugins.map((plugin) => plugin.pluginId),
+  ]);
+  const configuredChannelOwnerPluginIds = collectEffectiveConfiguredChannelOwnerPluginIds({
+    cfg: params.cfg,
+    env,
+    snapshot,
+    configuredChannelIds: params.channelIds,
+  });
+  const bundledPluginsById = new Map<string, BundledPluginPackageDescriptor>([
+    ...snapshot.plugins
+      .filter((plugin) => plugin.origin === "bundled")
+      .map((plugin) => [plugin.id, plugin] as const),
+    ...currentBundledPlugins.map(
+      (plugin) =>
+        [
+          plugin.pluginId,
+          {
+            packageName: plugin.packageName,
+          },
+        ] as const,
+    ),
+  ]);
+  const configuredPluginIdsWithStaleDescriptors =
+    collectConfiguredPluginIdsWithMissingChannelConfigDescriptors({
+      snapshot,
+      configuredPluginIds: params.pluginIds,
+      configuredChannelIds: params.channelIds,
+    });
+  const records = params.baselineRecords ?? (await loadInstalledPluginIndexInstallRecords({ env }));
+  const updateChannel = resolveRegistryUpdateChannel({
+    configChannel: normalizeUpdateChannel(params.cfg.update?.channel),
+    currentVersion: VERSION,
+  });
+  const installedPluginIdsWithRepairablePackageDiagnostics =
+    collectInstalledPluginIdsWithRepairablePackageDiagnostics({
+      snapshot,
+      installRecords: records,
+    });
+  const installedPluginIdsWithStaleVersionBoundRuntimePackages =
+    collectInstalledPluginIdsWithStaleVersionBoundRuntimePackages({
+      snapshot,
+      installRecords: records,
+      configuredPluginIds: params.pluginIds,
+      updateChannel,
+    });
+  const installedPluginIdsWithRepairablePackages = new Set([
+    ...installedPluginIdsWithRepairablePackageDiagnostics,
+    ...installedPluginIdsWithStaleVersionBoundRuntimePackages,
+  ]);
+  const officialReplacementInstallCandidates = collectOfficialReplacementInstallCandidates({
+    cfg: params.cfg,
+    env,
+    repairablePluginIds: installedPluginIdsWithRepairablePackages,
+    configuredPluginIds: params.pluginIds,
+    configuredChannelIds: params.channelIds,
+    configuredChannelOwnerPluginIds,
+    blockedPluginIds: params.blockedPluginIds,
+  });
+  const officialReplacementPluginIds = new Set(officialReplacementInstallCandidates.keys());
+  const changes: string[] = [];
+  const notices: string[] = [];
+  const warnings: string[] = [];
+  const deferredRepairDetails: string[] = [];
+  const failedPluginIds = new Set<string>();
+  const repairedPluginIds = new Set<string>();
+  const deferredPluginIds = new Set<string>();
+  const preferNpmInstalls = isLegacyPackageUpdateDoctorPass(env);
+  let nextRecords = records;
+
+  for (const [pluginId, record] of Object.entries(records)) {
+    const bundled = bundledPluginsById.get(pluginId);
+    if (!bundled || !recordMatchesBundledPackage(record, bundled)) {
+      continue;
+    }
+    if (nextRecords === records) {
+      nextRecords = { ...records };
+    }
+    delete nextRecords[pluginId];
+    changes.push(`Removed stale managed install record for bundled plugin "${pluginId}".`);
+  }
+
+  if (shouldDeferConfiguredPluginInstallRepair(env)) {
+    const updateDeferredPluginIds = collectUpdateDeferredPluginIds({
+      cfg: params.cfg,
+      env,
+      configuredPluginIds: params.pluginIds,
+      configuredChannelIds: params.channelIds,
+      configuredChannelOwnerPluginIds,
+      blockedPluginIds: params.blockedPluginIds,
+    });
+    for (const pluginId of updateDeferredPluginIds) {
+      deferredPluginIds.add(pluginId);
+      const record = nextRecords[pluginId];
+      if (!record || !isInstalledRecordMissingOnDisk(record, env)) {
+        continue;
+      }
+      const detail = `Skipped package-manager repair for configured plugin "${pluginId}" during package update; rerun "openclaw doctor --fix" after the update completes.`;
+      changes.push(detail);
+      deferredRepairDetails.push(detail);
+    }
+  }
+
+  const missingRecordedPluginIds = Object.keys(records).filter(
+    (pluginId) =>
+      !deferredPluginIds.has(pluginId) &&
+      !officialReplacementPluginIds.has(pluginId) &&
+      Object.hasOwn(nextRecords, pluginId) &&
+      !bundledPluginsById.has(pluginId) &&
+      ((params.pluginIds.has(pluginId) &&
+        (!knownIds.has(pluginId) || isInstalledRecordMissingOnDisk(nextRecords[pluginId], env))) ||
+        configuredPluginIdsWithStaleDescriptors.has(pluginId) ||
+        installedPluginIdsWithRepairablePackages.has(pluginId)),
+  );
+
+  if (missingRecordedPluginIds.length > 0) {
+    for (const pluginId of missingRecordedPluginIds) {
+      const record = nextRecords[pluginId];
+      if (!record) {
+        continue;
+      }
+      const forced = forceNpmInstallRecordRepair(record);
+      if (forced !== record) {
+        if (nextRecords === records) {
+          nextRecords = { ...records };
+        }
+        nextRecords[pluginId] = forced;
+      }
+    }
+    const updateResult = await updateNpmInstalledPlugins({
+      config: {
+        ...params.cfg,
+        plugins: {
+          ...params.cfg.plugins,
+          installs: nextRecords,
+        },
+      },
+      pluginIds: missingRecordedPluginIds,
+      updateChannel,
+      coreVersion: resolveCompatibilityHostVersion(env),
+      logger: {
+        terminalLinks: false,
+        warn: (message) => {
+          if (isClawHubReviewNotice(message)) {
+            notices.push(stripAnsi(message));
+            return;
+          }
+          warnings.push(message);
+        },
+        error: (message) => warnings.push(message),
+      },
+      ...(params.acknowledgeClawHubRisk ? { acknowledgeClawHubRisk: true } : {}),
+      ...(params.onClawHubRisk ? { onClawHubRisk: params.onClawHubRisk } : {}),
+    });
+    for (const outcome of updateResult.outcomes) {
+      if (outcome.status === "updated" || outcome.status === "unchanged") {
+        repairedPluginIds.add(outcome.pluginId);
+        changes.push(
+          installedPluginIdsWithStaleVersionBoundRuntimePackages.has(outcome.pluginId)
+            ? `Refreshed stale configured plugin "${outcome.pluginId}".`
+            : installedPluginIdsWithRepairablePackageDiagnostics.has(outcome.pluginId)
+              ? `Repaired broken installed plugin "${outcome.pluginId}".`
+              : `Repaired missing configured plugin "${outcome.pluginId}".`,
+        );
+      } else if (outcome.status === "error") {
+        warnings.push(outcome.message);
+        failedPluginIds.add(outcome.pluginId);
+      } else if (isActionableClawHubSkippedOutcome(outcome)) {
+        warnings.push(
+          appendClawHubRiskAcknowledgementGuidance({
+            message: outcome.message,
+            spec: recordClawHubInstallSpec(nextRecords[outcome.pluginId]),
+          }),
+        );
+        failedPluginIds.add(outcome.pluginId);
+      }
+    }
+    nextRecords = updateResult.config.plugins?.installs ?? nextRecords;
+  }
+
+  const missingPluginIds = new Set(
+    [...params.pluginIds].filter((pluginId) => {
+      if (deferredPluginIds.has(pluginId)) {
+        return false;
+      }
+      const hasRecord = Object.hasOwn(nextRecords, pluginId);
+      return (
+        (!knownIds.has(pluginId) && !hasRecord && !bundledPluginsById.has(pluginId)) ||
+        (hasRecord &&
+          !bundledPluginsById.has(pluginId) &&
+          isInstalledRecordMissingOnDisk(nextRecords[pluginId], env))
+      );
+    }),
+  );
+  const installCandidatePluginIds = new Set([...missingPluginIds, ...officialReplacementPluginIds]);
+  for (const candidate of collectDownloadableInstallCandidates({
+    cfg: params.cfg,
+    env,
+    missingPluginIds: installCandidatePluginIds,
+    configuredPluginIds: params.pluginIds,
+    configuredChannelIds: params.channelIds,
+    configuredChannelOwnerPluginIds,
+    blockedPluginIds:
+      deferredPluginIds.size > 0
+        ? new Set([...(params.blockedPluginIds ?? []), ...deferredPluginIds])
+        : params.blockedPluginIds,
+  })) {
+    if (bundledPluginsById.has(candidate.pluginId)) {
+      continue;
+    }
+    const shouldReplaceBrokenOfficialInstall = officialReplacementPluginIds.has(candidate.pluginId);
+    if (shouldReplaceBrokenOfficialInstall && !candidate.trustedSourceLinkedOfficialInstall) {
+      continue;
+    }
+    const record = nextRecords[candidate.pluginId];
+    if (
+      shouldReplaceBrokenOfficialInstall &&
+      !isTrustedOfficialInstallRecordForCandidate({ record, candidate })
+    ) {
+      continue;
+    }
+    const hasUsableRecord =
+      Object.hasOwn(nextRecords, candidate.pluginId) &&
+      !isInstalledRecordMissingOnDisk(nextRecords[candidate.pluginId], env);
+    if (
+      !shouldReplaceBrokenOfficialInstall &&
+      knownIds.has(candidate.pluginId) &&
+      hasUsableRecord
+    ) {
+      continue;
+    }
+    if (!shouldReplaceBrokenOfficialInstall && hasUsableRecord) {
+      continue;
+    }
+    const removalPath = shouldReplaceBrokenOfficialInstall
+      ? resolveSafeBrokenOfficialInstallRemovalPath({
+          pluginId: candidate.pluginId,
+          candidate,
+          record,
+          env,
+        })
+      : null;
+    const previousRecords = nextRecords;
+    const installed = await installCandidate({
+      candidate,
+      config: params.cfg,
+      records: nextRecords,
+      env,
+      updateChannel,
+      mode: shouldReplaceBrokenOfficialInstall ? "update" : "install",
+      preferNpm: preferNpmInstalls,
+      ...(installedPluginIdsWithStaleVersionBoundRuntimePackages.has(candidate.pluginId)
+        ? { repairReason: "stale-version-bound-runtime" as const }
+        : {}),
+      ...(params.acknowledgeClawHubRisk ? { acknowledgeClawHubRisk: true } : {}),
+      ...(params.onClawHubRisk ? { onClawHubRisk: params.onClawHubRisk } : {}),
+    });
+    if (shouldReplaceBrokenOfficialInstall) {
+      const installedRecord = installed.records[candidate.pluginId];
+      const replacementSucceeded = installed.records !== previousRecords;
+      if (
+        replacementSucceeded &&
+        removalPath &&
+        (!installedRecord?.installPath ||
+          !pathsEqual(resolveUserPath(installedRecord.installPath, env), removalPath))
+      ) {
+        try {
+          await rm(removalPath, { recursive: true, force: true });
+        } catch (error) {
+          warnings.push(
+            `Failed to remove broken installed plugin "${candidate.pluginId}" at ${removalPath}: ${String(error)}`,
+          );
+        }
+      }
+    }
+    nextRecords = installed.records;
+    changes.push(...installed.changes);
+    notices.push(...installed.notices);
+    warnings.push(...installed.warnings);
+    if (!installed.failedPluginId && installed.records[candidate.pluginId]) {
+      repairedPluginIds.add(candidate.pluginId);
+    }
+    if (installed.failedPluginId) {
+      failedPluginIds.add(installed.failedPluginId);
+    }
+  }
+
+  if (nextRecords !== records) {
+    await writePersistedInstalledPluginIndexInstallRecords(nextRecords, { env });
+  } else if (params.baselineRecords) {
+    // The caller seeded us from in-memory state that may not yet have been
+    // persisted (e.g. earlier sync/npm record mutations). Even if repair
+    // itself made no further changes, persist the baseline so the disk
+    // matches what we are about to return — otherwise the next reader gets
+    // a stale snapshot.
+    await writePersistedInstalledPluginIndexInstallRecords(nextRecords, { env });
+  }
+  return {
+    changes,
+    warnings,
+    ...(notices.length > 0 ? { notices } : {}),
+    ...(deferredRepairDetails.length > 0 ? { deferredRepairDetails } : {}),
+    ...(repairedPluginIds.size > 0
+      ? {
+          repairedPluginIds: [...repairedPluginIds].toSorted((left, right) =>
+            left.localeCompare(right),
+          ),
+        }
+      : {}),
+    ...(failedPluginIds.size > 0
+      ? {
+          failedPluginIds: [...failedPluginIds].toSorted((left, right) =>
+            left.localeCompare(right),
+          ),
+        }
+      : {}),
+    records: nextRecords,
+  };
+}
